+++ framework-system/framework-management/src/test/java/uk/gov/justice/services/management/ping/handler/PingHandlerTest.java
@@ -2,6 +2,7 @@ package uk.gov.justice.services.management.ping.handler;
 
 import static java.time.ZoneOffset.UTC;
 import static java.time.ZonedDateTime.of;
+import static java.util.UUID.randomUUID;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -9,6 +10,7 @@ import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.jmx.api.command.PingCommand;
 
 import java.time.ZonedDateTime;
+import java.util.UUID;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -16,7 +18,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.slf4j.Logger;
-
 
 @RunWith(MockitoJUnitRunner.class)
 public class PingHandlerTest {
@@ -33,11 +34,12 @@ public class PingHandlerTest {
     @Test
     public void shouldLogAPingMessage() throws Exception {
 
+
         final ZonedDateTime now = of(2019, 11, 7, 15, 22, 25, 0, UTC);
 
         when(clock.now()).thenReturn(now);
 
-        pingHandler.ping(new PingCommand());
+        pingHandler.ping(new PingCommand(), randomUUID());
 
         verify(logger).info("********** Received system command 'PING' at 2019-11-07T15:22:25Z **********");
     }