+++ jmx/jmx-command-handling/src/main/java/uk/gov/justice/services/jmx/runner/SystemCommandRunner.java
@@ -8,6 +8,8 @@ import uk.gov.justice.services.jmx.api.SystemCommandFailedException;
 import uk.gov.justice.services.jmx.api.command.SystemCommand;
 import uk.gov.justice.services.jmx.command.SystemCommandStore;
 
+import java.util.UUID;
+
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 
@@ -30,10 +32,10 @@ public class SystemCommandRunner {
     }
 
     @Transactional(NEVER)
-    public void run(final SystemCommand systemCommand) {
+    public void run(final SystemCommand systemCommand, final UUID commandId) {
 
         try {
-            systemCommandStore.findCommandProxy(systemCommand).invokeCommand(systemCommand);
+            systemCommandStore.findCommandProxy(systemCommand).invokeCommand(systemCommand, commandId);
         } catch (final Throwable e) {
             final String message = format("Failed to run System Command '%s'", systemCommand.getName());
             logger.error(message, e);